Change way highlight.js is imported
`highlight.js` is using *CommonJS* style modules, so *Rollup* can't
do tree shaking and the bundle was big, because installing it via
`npm` installs all the languages available.
Using the official custom builder of highlight.js seemed to be the
solution, but it won't play ball and I couldn't get it correctly
imported, so my remaining solutions were:
1. import the module with `npm` and manually edit it in `npm_modules`
   making it only export the desired modules.
2. import the module with `npm` and have a file substitute of the `index.js`
   of the module, in an static location, that wires and imports the proper
   files from `npm_modules`.

I chose this second way to do it.