Add some buttons for give format to text
Add buttons for inserting/formating text into strong, em and
strong, ordinated lists, unordinated lists, horizontal rules and
blockquotes.

If any text it selected, it'll be wrapped with the proper format
(or deleted and replaced with the format, in case it is an
horizontal rule). If not, a default text preformated with the
chosen format will be inserted.

I extracted this functionality to a module apart, in
`src/utils/`, so the components are a little more clean.