Improve text formatting buttons & little changes
Now, when we use a text formatting button, the default/affected
text keeps selected, for easy editing it.

I plan to extend the behaviour of the buttons in the future:
the idea is to allow formatting/unformatting a string clicking
multiple times the proper button, without deselecting the text.
This would help specially to have a heading insertion button.