Add button for image insertion
TODO: image insertion in the preview is a bit more complicated than
this. The main problem is that... an image can be pretty large and
this messes up with the scroll sync feature. We need a way to
calculate where the image is/their height and don't have it into account
for the scroll relative position calculation. Or it will be totally
unreliable.