First pass to RWD
For now, when the screen is small enought (800px) the app only shows
one view (editor or preview) at a time and also shows an extra button.
This extra button will serve to change between views and to indicate the
actual view. Later this change will can be done also with swipe gesture.
Even so, for now, this button is only a placeholder and not functional.