Make top bar menu responsive
The idea is to allow the menu to handle a wide number/size of items,
regardless of the screen size. To achieve that, we opted for an
horizontal scrolling menu. In touch screens, we can swipe the menu
as usual. In non touch capable screens (desktop), we can use the
mouse wheel to scroll the menu (or tab key, for focus the buttons).
Scrollbar is hidden, overlaped by MarkdownGroup component.

TODO: when the scrollbar is added, there is a «jump» in the buttons
position. Try to fix this later.